feat: converted from page router to app router, and made all page.tsx files in app dir to be pure server components. moved all 'use client' to component dir
