--- next.config.js
@@ -1,16 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-	reactStrictMode: true,
-	swcMinify: true,
-	compiler: {
-		styledComponents: true,
-	},
-	env: {
-		API_URL: process.env.API_URL,
-	},
-	eslint: {
-		ignoreDuringBuilds: true,
-	},
-};
-
-module.exports = nextConfig;
