+++ src/app/layout.tsx
@@ -1,0 +1,29 @@
+import AOSInit from "components/AOSInit";
+import ApolloProviderWrapper from "components/apolloProvider";
+import NProgress from "components/NProgress";
+// import StyledComponentsRegistry from "lib/styledRegistry";
+import "@fortawesome/fontawesome-free/css/all.min.css";
+import "../styles/index.scss";
+import "../styles/custom.scss";
+
+export const metadata = {
+	title: "Next.js",
+	description: "Generated by Next.js",
+};
+
+export default function RootLayout({ children }: { children: React.ReactNode }) {
+	return (
+		<html lang="en">
+			{/* <body>{children}</body> */}
+			<body>
+				<ApolloProviderWrapper>
+					{/* <StyledComponentsRegistry> */}
+					<AOSInit />
+					<NProgress />
+					{children}
+					{/* </StyledComponentsRegistry> */}
+				</ApolloProviderWrapper>
+			</body>
+		</html>
+	);
+}
