+++ src/components/home/BibleVersComp.tsx
@@ -1,3 +1,4 @@
+"use client";
 import React from "react";
 import { Autoplay, Navigation, Pagination } from "swiper";
 import { Swiper, SwiperSlide } from "swiper/react";
