+++ src/lib/apolloClient.ts
@@ -1,12 +1,12 @@
-import { ApolloClient, createHttpLink, InMemoryCache, NormalizedCacheObject } from '@apollo/client';
-import Cookies from 'js-cookie';
-import { NextPageContext } from 'next';
-import { useMemo } from 'react';
-import { TOKEN_NAME } from 'utils/constants';
-import { getTokenCookie } from '../utils/cookieUtils';
+import { ApolloClient, createHttpLink, InMemoryCache, NormalizedCacheObject } from "@apollo/client";
+import Cookies from "js-cookie";
+import { NextPageContext } from "next";
+import { useMemo } from "react";
+import { TOKEN_NAME } from "utils/constants";
+import { getTokenCookie } from "utils/cookieUtils";
 
 const API_URL = process.env.API_URL;
-const uri = API_URL + '/graphql';
+const uri = API_URL + "/graphql";
 let apolloClient: ApolloClient<unknown>;
 
 const token = Cookies.get(TOKEN_NAME);
@@ -14,19 +14,19 @@ const token = Cookies.get(TOKEN_NAME);
 export const apollo = new ApolloClient({
 	uri,
 	cache: new InMemoryCache(),
-	credentials: 'include',
+	credentials: "include",
 	headers: {
-		Authorization: token || ' ',
+		Authorization: token || " ",
 	},
 });
 
 const createLink = (initialState: unknown, token: string) => {
-	const cookie = typeof window !== 'undefined' ? Cookies.get(TOKEN_NAME) : token;
+	const cookie = typeof window !== "undefined" ? Cookies.get(TOKEN_NAME) : token;
 
 	const httpLink = createHttpLink({
 		uri,
 		fetch,
-		credentials: 'include',
+		credentials: "include",
 		headers: {
 			cookie,
 			Authorization: token,
@@ -34,7 +34,7 @@ const createLink = (initialState: unknown, token: string) => {
 	});
 
 	return new ApolloClient({
-		connectToDevTools: typeof window !== 'undefined',
+		connectToDevTools: typeof window !== "undefined",
 		ssrMode: typeof window === undefined,
 		link: httpLink,
 		cache: new InMemoryCache({ addTypename: false }).restore(initialState as NormalizedCacheObject),
@@ -47,9 +47,9 @@ export const initializeApollo = (
 ): ApolloClient<unknown> => {
 	const cookie = getTokenCookie(ctx?.req);
 
-	if (typeof window === undefined) return createLink(initialState, cookie || '');
+	if (typeof window === undefined) return createLink(initialState, cookie || "");
 	if (!apolloClient) {
-		apolloClient = createLink(initialState, cookie || '');
+		apolloClient = createLink(initialState, cookie || "");
 	}
 
 	return apolloClient;
