--- src/lib/styledRegistry.tsx
@@ -1,24 +1,0 @@
-'use client';
-
-import React, { useState } from 'react';
-import { useServerInsertedHTML } from 'next/navigation';
-import { ServerStyleSheet, StyleSheetManager } from 'styled-components';
-
-export default function StyledComponentsRegistry({ children }: { children: React.ReactNode }) {
-	// Only create stylesheet once with lazy initial state
-	// x-ref: https://reactjs.org/docs/hooks-reference.html#lazy-initial-state
-	const [styledComponentsStyleSheet] = useState(() => new ServerStyleSheet());
-
-	useServerInsertedHTML(() => {
-		const styles = styledComponentsStyleSheet.getStyleElement();
-		styledComponentsStyleSheet.instance.seal();
-		styledComponentsStyleSheet.seal();
-		return <>{styles}</>;
-	});
-
-	if (typeof window !== 'undefined') return <>{children}</>;
-
-	return (
-		<StyleSheetManager sheet={styledComponentsStyleSheet.instance}>{children as React.ReactElement}</StyleSheetManager>
-	);
-}
